Updated delete books to most recent main commit
