+++ app/__init__.py
@@ -2,22 +2,19 @@ from flask import Flask, Blueprint
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 from flask_login import LoginManager
-from config import Config
 from .errors import register_error_handlers
 
-bp = Blueprint('main', __name__)
 db = SQLAlchemy()
-migrate = Migrate()
-login_manager = LoginManager()   
+login_manager = LoginManager()
 
-def create_app():
-    global app
+def create_app(config):
     app = Flask(__name__)
-    app.config.from_object(Config)
+    app.config.from_object(config)
+
+    from app.blueprints import bp
+    app.register_blueprint(bp)
 
     db.init_app(app)
-    migrate.init_app(app, db)
-
     login_manager.init_app(app)
     login_manager.login_view = 'main.login'
 
@@ -27,8 +24,8 @@ def create_app():
     def load_user(user_id):
         return User.query.get(int(user_id))
 
-    from app.routes import bp as main_bp
-    app.register_blueprint(main_bp)
+    # from app import routes 
+    # from app import controllers 
 
     register_error_handlers(app)
