+++ app/templates/library.html
@@ -3,7 +3,9 @@
 {% block title %}Library - BookTracker{% endblock %}
 
 {% block head %}
+
 <link rel="stylesheet" href="{{ url_for('static', filename='css/library.css') }}">
+
 {% endblock %}
 
 {% block content %}
