+++ requirements.txt
@@ -3,3 +3,4 @@ Flask-WTF==1.2.2
 Flask-Migrate==4.1.0
 Flask-SQLAlchemy==3.1.1
 Flask-Login==0.6.3
+selenium==4.32.0
